pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baulkandcastle"
version = "1.0.0"
description = "Property tracking and valuation tool for Baulkham Hills & Castle Hill"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Antigravity (for Goran)"}
]
keywords = ["property", "real-estate", "scraper", "valuation", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]

dependencies = [
    # Web scraping with browser automation
    "crawl4ai==0.7.0",
    # HTML parsing
    "beautifulsoup4==4.12.3",
    # Data processing
    "pandas==2.2.0",
    "numpy==1.26.3",
    # Machine learning
    "scikit-learn==1.4.0",
    "xgboost==2.0.3",
    "joblib==1.3.2",
    # API server
    "flask==3.0.0",
    "flask-cors==4.0.0",
    # Browser automation (for Domain estimator)
    "playwright>=1.49.0",
    # Environment variables
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "pytest-asyncio==0.23.3",
    "black==23.12.1",
    "isort==5.13.2",
    "flake8==7.0.0",
    "mypy==1.8.0",
    "pre-commit==3.6.0",
    "types-requests==2.31.0.20240106",
    "pandas-stubs==2.1.4.231227",
]
test = [
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "pytest-asyncio==0.23.3",
]

[project.scripts]
# CLI entry points
baulkandcastle-scrape = "baulkandcastle.cli.scraper:main"
baulkandcastle-api = "baulkandcastle.cli.api_server:main"
baulkandcastle-train = "baulkandcastle.cli.train_model:main"
baulkandcastle-predict = "baulkandcastle.cli.predict:main"
baulkandcastle-estimate = "baulkandcastle.cli.estimate_for_sale:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
baulkandcastle = ["py.typed"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | \.pytest_cache
    | migrations
    | \.mypy_cache
    | frontend/node_modules
    | frontend/dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["baulkandcastle"]
skip = [".git", ".venv", "__pycache__", "migrations", "frontend"]
src_paths = ["src", "tests"]

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
    "unit: marks unit tests",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/baulkandcastle"]
branch = true
omit = [
    "*/__pycache__/*",
    "*/tests/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 60
show_missing = true

# mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "migrations/",
    "frontend/",
    ".venv/",
]

[[tool.mypy.overrides]]
module = [
    "crawl4ai.*",
    "xgboost.*",
    "sklearn.*",
    "flask_cors.*",
]
ignore_missing_imports = true
